movie-actor-ranking-api/pyproject.toml
[project]
name = "movie-actor-ranking-api"
version = "0.1.0"
description = "Movie Actor Ranking API"
requires-python = ">=3.12"
dependencies = [
    "aiocache>=0.12.3",
    "beautifulsoup4>=4.14.3",
    "cinemagoer>=2023.5.1",
    "fastapi>=0.135.2",
    "fuzzywuzzy>=0.18.0",
    "marshmallow>=4.2.3",
    "matplotlib>=3.10.8",
    "nltk>=3.9.4",
    "numpy>=2.4.4",
    "pandas>=3.0.1",
    "sqlmodel>=0.0.37",
    "asyncpg>=0.31.0",
    "python-dotenv>=1.2.2",
    "python-levenshtein>=0.27.3",
    "scikit-learn>=1.8.0",
    "selenium>=4.41.0",
    "slowapi>=0.1.9",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "transformers>=5.4.0",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
